+++ tests/schemas.py
@@ -24,7 +24,7 @@ class Schema_01(Schema):
 class Schema_02(Schema):
     int_attr = Int("int_attr", default=123)
     str_attr = Str("str_attr", default="test")
-    list_attr = List("list_attr", NestedSchema_01())
+    list_attr = List(item=NestedSchema_01())
     str_attr_null = Str(default=None)
     nested = NestedSchema_01("nested")
 
@@ -37,7 +37,7 @@ class Schema_03(Schema):
     float_attr = Float("float_attr", default=1.23, help="a float attribute")
     list_attr_int = List("list_attr_int", Int("list_attr_int_item"))
     list_attr_str = List("list_attr_str", Str("list_attr_str_item"))
-    list_attr_schema = List("list_attr_schema", NestedSchema_01())
+    list_attr_schema = List(name="list_attr_schema", item=NestedSchema_01())
     nested = NestedSchema_02("nested")
     int_attr_disabled = Int("int_attr_disabled", cli=False)
 
@@ -46,7 +46,7 @@ class Schema_04(Schema):
     str_attr = Str("str_attr", default="test")
     str_attr_null = Str(default=None)
     list_attr = List("list_attr", NestedSchema_01())
-    list_attr_w_default = List("list_attr_w_default", Int(), default=[1,2,3])
+    list_attr_w_default = List("list_attr_w_default", item=Int(), default=[1,2,3])
     nested = NestedSchema_02("nested")
 
 class Schema_05(Schema):