+++ confu/exceptions.py
@@ -1,4 +1,8 @@
 class SoftDependencyError(ImportError):
+    """
+    Raised when a feature requires a dependency that is missing
+    """
+
     def __init__(self, dep_name):
         super(SoftDependencyError, self).__init__(
             "To use this feature this dependency is required: {}".format(dep_name)
@@ -6,7 +10,20 @@ class SoftDependencyError(ImportError):
 
 
 class ValidationErrorBase(ValueError):
+    """
+    Config validation error interface
+    """
+
     def __init__(self, attribute, path, value, reason):
+        """
+        **Arguments**
+
+        - attribute (`Attribute`): confu attribute instance
+        - path (`list`): attribute path
+        - value (`mixed`): value that caused the validation error
+        - reason (`str`): human readable reason message for validation error
+        """
+
         msg = "{}: {}".format(path, reason)
         self.details = {
             "path": path,
@@ -18,6 +35,9 @@ class ValidationErrorBase(ValueError):
 
     @property
     def pretty(self):
+        """
+        pretty formatted error message
+        """
         return "{}: {}".format(
             ".".join([str(i) for i in self.details["path"]]), self.details["reason"]
         )
@@ -34,13 +54,25 @@ class ValidationErrorBase(ValueError):
 
 
 class ValidationWarning(ValidationErrorBase):
+    """
+    Config validation warning
+    """
+
     pass
 
 
 class ValidationError(ValidationErrorBase):
+    """
+    Config validation error
+    """
+
     pass
 
 
 class ApplyDefaultError(ValidationErrorBase):
+    """
+    Raised when an exception occured during apply_defaults
+    """
+
     def __str__(self):
         return self.pretty