+++ confu/schema/inet.py
@@ -1,3 +1,13 @@
+"""
+Attributes that deal with networking specific values such as emails, urls and ip addresses
+
+These can be imported directly from `confu.schema`
+
+## Requirements
+
+- `ipaddress` for ip address validation
+"""
+
 import re
 
 try:
@@ -100,22 +110,22 @@ class IpAddress(Str):
 
         **Keyword Arguments**
 
-        - `name` (str): describes the attribute name, if not specified
+        - name (`str`): describes the attribute name, if not specified
           explicitly will be set through the schema that instantiates
           the attribute.
-        - `protocol` (int): ip version, can be 4, 6 or None - if it is none
+        - protocol (`int`): ip version, can be 4, 6 or None - if it is none
           the attribute can hold either a v4 or a v6 IP address.
-        - `default` (mixed): the default value of this attribute. Once a default
+        - default (`mixed`): the default value of this attribute. Once a default
           value is set, schema validation will no longer raise a
           validation error if the attribute is missing from the
           configuration.
-        - `choices` (list): if specified on values in this list may be set
+        - choices (`list`): if specified on values in this list may be set
           for this attribute
-        - `help` (str): help description
-        - `cli` (bool=True): enable CLI support for this attribute
-        - `deprecated` (str): version id of when this attribute will be deprecated
-        - `added` (str): version id of when this attribute was added to the schema
-        - `removed` (str): version id of when this attribute will be removed
+        - help (`str`): help description
+        - cli (`bool=True`): enable CLI support for this attribute
+        - deprecated (`str`): version id of when this attribute will be deprecated
+        - added (`str`): version id of when this attribute was added to the schema
+        - removed (`str`): version id of when this attribute will be removed
         """
 
         super(IpAddress, self).__init__(name=name, **kwargs)